dt-bindings: iio: dac: Add adi,ltc2664.yaml
The LTC2664 is a family of 4-channel, 12/16-bit +/-10V digital-to-analog
converters with integrated precision references. They are guaranteed
monotonic and have built-in rail-to-rail output buffers.
These SoftSpan DACs offer five output ranges up to +/-10V.
The range of each channel is independently programmable, or the part can
be hardware-configured for operation in a fixed range.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>